falta ver q lea el vacio de comentarios

+++ main.py
@@ -20,7 +20,7 @@ class Aplicacion():
 		self.frame_et.pack(side = LEFT, fill = BOTH)
 		self.frame_cam.pack(side = LEFT, fill = BOTH, expand = 1)
 
-		# Declara variables de control
+		#VARIABLES DE CONTROL
 		self.id = IntVar()
 		self.nombre = StringVar()
 		self.clave = StringVar()
@@ -33,15 +33,14 @@ class Aplicacion():
 					 \n4. Para actualizar una entrada, ejecuta el paso 3, cambia los datos deseados y presiona act.
 					 \n5. Para borrar una entrada, ejecuta el paso 3 y presiona en borrar.
 					'''
-
-		# Declara widgets de la self.raiz
-        # Se incluye el widget de tipo Button 'Crear' 'Leer' 'Actualizar' 'Borrar' 
-		self.et_id = tkinter.Label(self.frame_et, text = "ID")
-		self.et_nombre = tkinter.Label(self.frame_et, text = "Nombre")
-		self.et_apellido = tkinter.Label(self.frame_et, text = "Apellido")
-		self.et_clave = tkinter.Label(self.frame_et, text = "Clave")
-		self.et_direccion = tkinter.Label(self.frame_et, text = "Direccion")
-		self.et_comentarios = tkinter.Label(self.frame_et, text = "Comentarios")
+		self.conectado = False
+		#ELEMENTOS DE LA VENTANA
+		self.et_id = tkinter.Label(self.frame_et, text = "ID:")
+		self.et_nombre = tkinter.Label(self.frame_et, text = "Nombre:")
+		self.et_apellido = tkinter.Label(self.frame_et, text = "Apellido:")
+		self.et_clave = tkinter.Label(self.frame_et, text = "Clave:")
+		self.et_direccion = tkinter.Label(self.frame_et, text = "Direccion:")
+		self.et_comentarios = tkinter.Label(self.frame_et, text = "Comentarios:")
 		self.campo_id = tkinter.Entry(self.frame_cam, textvariable = self.id)
 		self.campo_id.delete(0, len(self.campo_id.get()))
 		self.campo_nombre = tkinter.Entry(self.frame_cam, textvariable = self.nombre)
@@ -52,15 +51,14 @@ class Aplicacion():
 		self.boton_crear = tkinter.Button(self.frame_bot, text = "Crear", command = self.crear_entrada)
 		self.boton_leer = tkinter.Button(self.frame_bot, text = "Leer", command = self.leer_entrada)
 		self.boton_act = tkinter.Button(self.frame_bot, text = "Actualizar", command = self.actualizar_entrada)
-		self.boton_borrar = tkinter.Button(self.frame_bot, text = "Borrar", command = self.borrar_entrada)
-		
+		self.boton_borrar = tkinter.Button(self.frame_bot, text = "Borrar", command = self.borrar_entrada)		
 		#MENU SUPERIOR
 		self.menu_bbdd = Menu(self.barra_menu, tearoff = 0)
 		self.menu_borrar = Menu(self.barra_menu, tearoff = 0)
 		self.menu_crud = Menu(self.barra_menu, tearoff = 0)
 		self.menu_ayuda = Menu(self.barra_menu, tearoff = 0)
 		self.barra_menu.add_cascade(label = "BBDD", menu = self.menu_bbdd)
-		self.menu_bbdd.add_command(label = "Conectar", command = lambda: self.conectar_base())
+		self.menu_bbdd.add_command(label = "Conectar", command = lambda: self.conectar_base(0))
 		self.menu_bbdd.add_command(label = "Salir", command = lambda: self.ventana_mensaje(True, ""))
 		self.barra_menu.add_cascade(label = "Borrar", menu = self.menu_borrar)
 		self.menu_borrar.add_command(label = "Borrar campos", command = lambda: self.borrar_campos())
@@ -90,22 +88,30 @@ class Aplicacion():
 		self.boton_act.pack(side = LEFT, fill = BOTH, expand = True, padx = 3)
 		self.boton_borrar.pack(side = LEFT, fill = BOTH, expand = True, padx = 3)
 		self.raiz.mainloop()
-		self.conexion.commit()
 			
 	#CREA TODAS LAS VENTANAS EMERGENTES
 	def ventana_mensaje(self, salir, mensaje):
 		if salir:
 			r = mb.askquestion("Salir", "¿Desea salir de la aplicacion?")
 			if r == "yes":
-				self.conexion.close()
+				if self.conectado:
+					self.conexion.close()
 				self.raiz.quit()
 		else:
-			mb.showinfo("BBDD", mensaje)
+			mb.showinfo("Base de datos", mensaje)
 
-	#SE CONECTA A LA BASE DE DATOS Y CREA LA TABLA
-	def conectar_base(self):
-		self.conexion = sqlite3.connect("Base_personas")
-		self.cursor = self.conexion.cursor()
+	#SE CONECTA A LA BASE Y GUARDA LOS DATOS
+	def conectar_base(self, opcion):
+		if (opcion == 0):
+			self.conexion = sqlite3.connect("Base_personas")
+			self.cursor = self.conexion.cursor()
+			self.crear_tabla()
+			self.conectado = True
+		else:
+			self.conexion.commit()
+
+	#CREA LA TABLA DE LA BASE DE DATOS
+	def crear_tabla(self):
 		try:
 			self.cursor.execute('''
 			CREATE TABLE registro_personas(
@@ -127,48 +133,71 @@ class Aplicacion():
 		self.campo_clave.delete(0, len(self.campo_clave.get()))
 		self.campo_apellido.delete(0, len(self.campo_apellido.get()))
 		self.campo_direccion.delete(0, len(self.campo_direccion.get()))
-		self.campo_comentarios.delete(1.0, END) 
+		self.campo_comentarios.delete(1.0, END)
+
+	#REVISA SI HAY DATOS VACIOS
+	def revisa_datos(self, lista):
+		vacio = False
+		for dato in lista:
+			if (dato == ""):
+				vacio = True
+		return vacio
 
 	#AGREGA UNA NUEVA ENTRADA AL REGISTRO
 	def crear_entrada(self):
 		lista_datos = (self.nombre.get(), self.clave.get(), self.apellido.get(), self.direccion.get(), self.campo_comentarios.get(1.0, END))
-		self.cursor.execute('INSERT INTO registro_personas VALUES (NULL, ?, ?, ?, ?, ?)', lista_datos)
-		self.borrar_campos()
-		self.ventana_mensaje(False, "Registro creado con exito")
+		if self.conectado:
+			if (not self.revisa_datos(lista_datos)):
+				self.cursor.execute('INSERT INTO registro_personas VALUES (NULL, ?, ?, ?, ?, ?)', lista_datos)
+				self.borrar_campos()
+				self.ventana_mensaje(False, "Registro creado con exito")
+				self.conectar_base(1)
+			else:
+				self.ventana_mensaje(False, "Error, hay campos vacios")
+		else:
+			self.ventana_mensaje(False, "Base de datos desconectada")
 
 	#MUESTRA LOS DATOS DE UNA ENTRADA INDICANDO NUMERO DE ID
 	def leer_entrada(self):
-		try:
-			self.cursor.execute('SELECT * FROM registro_personas WHERE ID=:num_id', {'num_id': self.id.get()})
-			lista_datos = self.cursor.fetchone()
-			self.borrar_campos()
-			self.campo_id.insert(0, lista_datos[0])
-			self.campo_nombre.insert(0, lista_datos[1])
-			self.campo_clave.insert(0, lista_datos[2])
-			self.campo_apellido.insert(0, lista_datos[3])
-			self.campo_direccion.insert(0, lista_datos[4])
-			self.campo_comentarios.insert(END, lista_datos[5]) 
-		except:
-			self.ventana_mensaje(False, "Ese registro no existe")
+		if self.conectado:
+			try:
+				self.cursor.execute('SELECT * FROM registro_personas WHERE ID=:num_id', {'num_id': self.id.get()})
+				lista_datos = self.cursor.fetchone()
+				self.borrar_campos()
+				self.campo_id.insert(0, lista_datos[0])
+				self.campo_nombre.insert(0, lista_datos[1])
+				self.campo_clave.insert(0, lista_datos[2])
+				self.campo_apellido.insert(0, lista_datos[3])
+				self.campo_direccion.insert(0, lista_datos[4])
+				self.campo_comentarios.insert(END, lista_datos[5]) 
+			except:
+				self.ventana_mensaje(False, "Ese registro no existe, ingrese un numero valido")
+		else:
+			self.ventana_mensaje(False, "Base de datos desconectada")
 
 	#ACTUALIZA DATOS DE UNA ENTRADA EXISTENTE
 	def actualizar_entrada(self):
-		lista_nuevos = [self.nombre.get(), self.clave.get(), self.apellido.get(), self.direccion.get(), self.campo_comentarios.get(1.0, END)]
-		self.cursor.execute('''
-							UPDATE registro_personas SET 
-							nombre=:nuevo_nombre, clave=:clave_nuevo, 
-							apellido=:apellido_nuevo, direccion=:direccion_nuevo,
-							comentarios=:comentarios_nuevos
-			 				WHERE ID=:num_id''', 
-			 				{'nuevo_nombre': lista_nuevos[0], 'clave_nuevo': lista_nuevos[1], 'apellido_nuevo': lista_nuevos[2],
-			 				'direccion_nuevo': lista_nuevos[3], 'comentarios_nuevos': lista_nuevos[4], 'num_id': self.id.get()})
-		self.ventana_mensaje(False, "Registro actualizado con exito")
+		if self.conectado:
+			lista_nuevos = (self.nombre.get(), self.clave.get(), self.apellido.get(), self.direccion.get(), self.campo_comentarios.get(1.0, END))
+			if (not self.revisa_datos(lista_nuevos)):
+				self.cursor.execute('''UPDATE registro_personas SET NOMBRE=?, CLAVE=?, APELLIDO=?, DIRECCION=?, COMENTARIOS=? 
+					 				   WHERE ID=''' + str(self.id.get()), lista_nuevos)
+				self.ventana_mensaje(False, "Registro actualizado con exito")
+				self.conectar_base(1)
+			else:
+				self.ventana_mensaje(False, "Error, hay campos vacios")
+		else:
+			self.ventana_mensaje(False, "Base de datos desconectada")			
 
 	#ELIMINA UNA ENTRADA DE LA BASE DE DATOS
 	def borrar_entrada(self):
-		self.cursor.execute('DELETE FROM registro_personas WHERE ID=:num_id', {'num_id': self.id.get()})
-		self.borrar_campos()
-		self.ventana_mensaje(False, "Registro borrado con exito")
+		if self.conectado:
+			self.cursor.execute('DELETE FROM registro_personas WHERE ID=:num_id', {'num_id': self.id.get()})
+			self.borrar_campos()
+			self.ventana_mensaje(False, "Registro borrado con exito")
+			self.conectar_base(1)
+		else:
+			self.ventana_mensaje(False, "Base de datos desconectada")
 
 def main():
     mi_app = Aplicacion()
